Move components of a game into ave.components
