+++ ave/components/__init__.py
@@ -1,0 +1,3 @@
+from .with_requirements import TextWithRequirements  # noqa: F401
+from .with_requirements import NameWithRequirements  # noqa: F401
+from .with_requirements import OptionWithRequirements  # noqa: F401
