+++ ave/game.py
@@ -1,10 +1,9 @@
 """Classes for running games."""
 
-from random import randrange
 from .exceptions import AVEGameOver, AVEWinner, AVEVersionError
 from .string_functions import more_unescape
 from .items import Number
-from .requirements import Satisfied
+from .components import TextWithRequirements, OptionWithRequirements
 from . import config
 
 
@@ -43,7 +42,7 @@ class Character:
         self.inventory = []
         self.numbers = {}
         for i in items.values():
-            if isinstance(i, Number):
+            if self.is_number(i):
                 self.numbers[i.id] = i.default.get_value(self)
         self.location = "start"
 
@@ -294,76 +293,3 @@ class Room:
         """
         return {i: o for i, o in enumerate(self.options)
                 if o.has_requirements(character)}
-
-
-class ThingWithRequirements:
-    """A thing that needs requirements to be satisfied to be shown."""
-
-    def __init__(self, items=[], needs=Satisfied()):
-        """Make the thing."""
-        self.items = items
-        self.needs = needs
-
-    def has_requirements(self, character):
-        """Check is requirements are satisfied."""
-        return self.needs.has(character)
-
-    def get_items(self, character):
-        """Give the character the items for and/or take items away."""
-        for item in self.items:
-            item.give(character)
-
-
-class TextWithRequirements(ThingWithRequirements):
-    """A line of test that needs requirements to be satisfied to be shown."""
-
-    def __init__(self, text, **kwargs):
-        """Make the thing."""
-        self.text = text
-        super().__init__(**kwargs)
-
-
-class OptionWithRequirements(ThingWithRequirements):
-    """An option that needs requirements to be satisfied to be shown."""
-
-    def __init__(self, text, destination, random=None, **kwargs):
-        """Make the thing."""
-        self.text = text
-        self.destination = destination
-        if random is None:
-            self.random = False
-        else:
-            self.random = True
-            self.probabilities = random
-        super().__init__(**kwargs)
-
-    def get_all_destinations(self):
-        """Get a list of all destinations that this options could lead to.
-
-        Randomness will lead to multiple destinations from a single option.
-        """
-        if self.random:
-            return self.destination
-        else:
-            return [self.destination]
-
-    def get_destination(self):
-        """Get the destination when this option is chosen."""
-        if self.random:
-            n = randrange(sum(self.probabilities))
-            total = 0
-            for d, i in zip(self.destination, self.probabilities):
-                total += i
-                if total > n:
-                    return d
-        else:
-            return self.destination
-
-
-class NameWithRequirements(ThingWithRequirements):
-    """The name of an item that needs requirements to be satisfied."""
-
-    def __init__(self, text, **kwargs):
-        """Make the thing."""
-        self.text = text
-        super().__init__(**kwargs)
